Updated many files to account for multiple instances
sortVec was changed to take in a parameter, also made it a more simple bubbleSort. moveThrough was changed to accept queues rather than use myQueue only. Also added an input pass to the function and polymorphed it. Polymorphed closeDoors to take in an int input. myElevator now is an array. main has 3 elevators now as well. And other things that I may have forgotten.